Add files via upload
had a problem with lab 3 where I was unable to edit the code for part 1 to accommodate part two, so I had to make a second one from scratch that did not have a magnetic field variable(since one was not required for part 2). Still have no idea why the plots become nonsense once I reduced H to zero.