Add files via upload
The program takes a bit of time to run, but as long as the variables are not to extreme(by themselves or in comparison to each other) and the time step and length are not to great, it shouldn't break any computers.
I learned a fair amount this semester and i'm grateful for it. My last two projects went poorly and this one will likely be so as well, but even so, I defiantly know a lot more then I did when I started this class, especially hands on experience!. Thank you for the help and I hope you have a great(and not so hot) summer.